+++ src/extensions/javascript-extension.ts
@@ -29,6 +29,25 @@ export default function (): Extension {
                         },
                     });
                 },
+                suggestions: [
+                    { suggestion: "template.path.path" },
+                    { suggestion: "template.path.parent" },
+                    { suggestion: "template.path.name" },
+                    { suggestion: "template.path.basename" },
+                    { suggestion: "template.path.extension" },
+                    {
+                        suggestion: "template.properties.$insertTo",
+                    },
+                    { suggestion: "template.properties.{key}" },
+                    // TODO: provide a way to add suggestion here
+                    ...["pochoir:form", "pochoir:date", "{import}"].map(
+                        (name) => ({
+                            suggestion: `await template.import("${name}")`,
+                            trigger: `template.import("${name}")`,
+                            display: `template.import("${name}")`,
+                        }),
+                    ),
+                ],
             });
         },
     };